Disable telemetry when validating config (#2)
This PR disables telemetry when validating the config. Without this, the first time precommit runs the hook in an interactive terminal, it will hang indefinitely. It is waiting for the user to make a choice on whether to send telemetry to CircleCI, but that log is not shown in the terminal.